requirements.txt
streamlit>=1.20.0
pandas>=2.0.0
pillow>=9.0.0
plotly>=5.0.0
openpyxl>=3.0.0
python-dateutil>=2.8.2
